requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.23.0
scipy>=1.9.0
librosa>=0.10.0
soundfile>=0.12.1
opencv-python>=4.7.0
Pillow>=9.4.0
dlib>=19.24.0
facenet-pytorch>=2.5.3
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.64.0
PyYAML>=6.0
tensorboard>=2.12.0
einops>=0.6.0
timm>=0.9.0
pingouin>=0.5.3
pytest>=7.2.0
